move revert state
